Refactor jdeps extension to prepare for K2 impl
